Works but still need to fix up a couple things
I'll modify the draw_grid() after everything else is working fine it that's alright.
WORKS:
hit_ship() works and draw() displays the current players ships and all attacks from both players. 
DOES NOT WORK:
In draw(), lines 78-85, I am not able to to loop through all the ships. (The commented parts). The uncommented lines display the ship, but it the wrong location when it is the right players turn.